+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds Tugboat QA configuration for PR preview deployments to replace Netlify.
 - Moved SCSS and CSS variable styles into a custom NYPL theme to extend Chakra's base theme.
 - Updates the `DatePicker`, `TextInput`, `Select`, and `Label` components to pass an optional `showOptReqLabel` prop to conditionally render "Required"/"Optional" in the label text.
+- Updates the `errorText` prop name to `invalidText`.
 
 ### Fixes